+++ examples/anyscale-v2-kitchensink/variables.tf
@@ -46,6 +46,27 @@ variable "s3_bucket_name" {
   type        = string
 }
 
+variable "anyscale_external_id" {
+  description = <<-EOF
+    (Required) A string that will be used for the IAM trust policy.
+    The trust policy for the control plane IAM role will be locked down to the provided external ID.
+    The external ID must start with the Organization ID (e.g. `org_1234567890abcdef`)
+
+    ex:
+    ```
+    anyscale_external_id = "org_1234567890abcdef-external-id-12345"
+    ```
+  EOF
+  type        = string
+  validation {
+    condition = (
+      length(var.anyscale_external_id) > 4 &&
+      substr(var.anyscale_external_id, 0, 4) == "org_"
+    )
+    error_message = "The anyscale_external_id value must start with \"org_\"."
+  }
+}
+
 # ------------------------------------------------------------------------------
 # OPTIONAL PARAMETERS
 # These variables have defaults, but may be overridden.
@@ -171,4 +192,10 @@ variable "security_group_enable_ssh_access" {
   EOT
   type        = bool
   default     = true
+}
+
+variable "anyscale_s3_force_destroy" {
+  description = "This is used to set the S3 force destroy value for testing purposes"
+  type        = bool
+  default     = false
 }