+++ examples/anyscale-v2-privatesubnets/variables.tf
@@ -22,28 +22,31 @@ variable "aws_region" {
   type        = string
 }
 
-# ------------------------------------------------------------------------------
-# OPTIONAL PARAMETERS
-# These variables have defaults, but may be overridden.
-# ------------------------------------------------------------------------------
-variable "anyscale_deploy_env" {
+variable "anyscale_external_id" {
   description = <<-EOF
-    (Optional) Anyscale deployment environment.
-    Used in resource names and tags.
+    (Required) A string that will be used for the IAM trust policy.
+    The trust policy for the control plane IAM role will be locked down to the provided external ID.
+    The external ID must start with the Organization ID (e.g. `org_1234567890abcdef`)
+
     ex:
     ```
-    anyscale_deploy_env = "production"
+    anyscale_external_id = "org_1234567890abcdef-external-id-12345"
     ```
   EOF
   type        = string
   validation {
     condition = (
-      var.anyscale_deploy_env == "production" || var.anyscale_deploy_env == "development" || var.anyscale_deploy_env == "test"
+      length(var.anyscale_external_id) > 4 &&
+      substr(var.anyscale_external_id, 0, 4) == "org_"
     )
-    error_message = "The anyscale_deploy_env only allows `production`, `test`, or `development`"
+    error_message = "The anyscale_external_id value must start with \"org_\"."
   }
-  default = "production"
 }
+
+# ------------------------------------------------------------------------------
+# OPTIONAL PARAMETERS
+# These variables have defaults, but may be overridden.
+# ------------------------------------------------------------------------------
 
 variable "anyscale_cloud_id" {
   description = <<-EOF
@@ -69,6 +72,8 @@ variable "anyscale_cloud_id" {
   }
 }
 
+
+
 variable "tags" {
   description = <<-EOF
     (Optional) A map of tags.
@@ -92,4 +97,10 @@ variable "s3_tag_value" {
   description = "This is used to set the S3 tag value for testing purposes"
   type        = string
   default     = "testing"
+}
+
+variable "anyscale_s3_force_destroy" {
+  description = "This is used to set the S3 force destroy value for testing purposes"
+  type        = bool
+  default     = false
 }